requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
captcha-0.7.1==0.7.1